move test func to test file
