+++ p2p/transport_test.go
@@ -9,12 +9,25 @@ import (
 	"time"
 
 	"github.com/tendermint/tendermint/crypto/ed25519"
+	"github.com/tendermint/tendermint/p2p/conn"
 )
 
 var defaultNodeName = "host_peer"
 
 func emptyNodeInfo() NodeInfo {
 	return DefaultNodeInfo{}
+}
+
+// newMultiplexTransport returns a tcp connected multiplexed peer
+// using the default MConnConfig. It's a convenience function used
+// for testing.
+func newMultiplexTransport(
+	nodeInfo NodeInfo,
+	nodeKey NodeKey,
+) *MultiplexTransport {
+	return NewMultiplexTransport(
+		nodeInfo, nodeKey, conn.DefaultMConnConfig(),
+	)
 }
 
 func TestTransportMultiplexConnFilter(t *testing.T) {
